check auth before accessing specific resources

+++ graphql-prisma/src/resolvers/Mutation.js
@@ -1,5 +1,5 @@
 import bcrypt from "bcryptjs";
-import jwt from "jsonwebtoken";
+import JWT from "jsonwebtoken";
 
 export default {
   async createUser(_parent, args, { prisma }, _info) {
@@ -18,7 +18,7 @@ export default {
 
     return {
       user,
-      token: jwt.sign({ userId: user.id }, "SECURE")
+      token: JWT.sign({ userId: user.id }, "SECURE")
     };
   },
 
@@ -31,48 +31,55 @@ export default {
 
     return {
       user,
-      token: jwt.sign({ userId: user.id }, "SECURE")
+      token: JWT.sign({ userId: user.id }, "SECURE")
     };
   },
 
-  async deleteUser(_parent, args, { prisma }, info) {
-    const userExists = await prisma.exists.User({ id: args.id });
-    if (!userExists) throw new Error("User doesn't exist");
-
-    return await prisma.mutation.deleteUser({ where: { id: args.id } }, info);
+  async deleteUser(_parent, _args, { prisma, userId }, info) {
+    if (!userId) throw new Error("Authentication required");
+    return await prisma.mutation.deleteUser({ where: { id: userId } }, info);
   },
 
-  async updateUser(_parent, { id, data }, { prisma }, info) {
-    const userExists = await prisma.exists.User({ id });
-    if (!userExists) throw new Error("User doesn't exist");
-
-    return await prisma.mutation.updateUser({ where: { id }, data }, info);
-  },
-
-  async createPost(
-    _parent,
-    { data: { title, body, published, author } },
-    { prisma },
-    info
-  ) {
-    const authorExists = await prisma.exists.User({ id: author });
-    if (!authorExists) throw new Error("Author doesn't exist");
-
-    return await prisma.mutation.createPost(
-      { data: { title, body, published, author: { connect: { id: author } } } },
+  async updateUser(_parent, { data }, { prisma, userId }, info) {
+    if (!userId) throw new Error("Authentication required");
+    return await prisma.mutation.updateUser(
+      { where: { id: userId }, data },
       info
     );
   },
 
-  async updatePost(_parent, { id, data }, { prisma }, info) {
-    const postExists = await prisma.exists.Post({ id });
+  async createPost(
+    _parent,
+    { data: { title, body, published } },
+    { prisma, userId },
+    info
+  ) {
+    if (!userId) throw new Error("Authentication required");
+    return await prisma.mutation.createPost(
+      { data: { title, body, published, author: { connect: { id: userId } } } },
+      info
+    );
+  },
+
+  async updatePost(_parent, { id, data }, { prisma, userId }, info) {
+    if (!userId) throw new Error("Authentication required");
+
+    const postExists = await prisma.exists.Post({
+      id,
+      author: { id: userId }
+    });
     if (!postExists) throw new Error("Post doesn't exist");
 
     return await prisma.mutation.updatePost({ where: { id }, data }, info);
   },
 
-  async deletePost(_parent, { id }, { prisma }, info) {
-    const postExists = await prisma.exists.Post({ id });
+  async deletePost(_parent, { id }, { prisma, userId }, info) {
+    if (!userId) throw new Error("Authentication required");
+
+    const postExists = await prisma.exists.Post({
+      id,
+      author: { id: userId }
+    });
     if (!postExists) throw new Error("Post doesn't exist");
 
     return await prisma.mutation.deletePost({ where: { id } }, info);
@@ -80,21 +87,20 @@ export default {
 
   async createComment(
     _parent,
-    { data: { text, post, author } },
-    { prisma },
+    { data: { text, post } },
+    { prisma, userId },
     info
   ) {
-    const postExists = await prisma.exists.Post({ id: post });
-    const authorExists = await prisma.exists.User({ id: author });
+    if (!userId) throw new Error("Authentication required");
 
-    if (!authorExists || !postExists)
-      throw new Error("Author Or Post does not exist!");
+    const postExists = await prisma.exists.Post({ id: post });
+    if (!postExists) throw new Error("Post does not exist!");
 
     return await prisma.mutation.createComment(
       {
         data: {
           text,
-          author: { connect: { id: author } },
+          author: { connect: { id: userId } },
           post: { connect: { id: post } }
         }
       },
@@ -102,15 +108,25 @@ export default {
     );
   },
 
-  async updateComment(_parent, { id, data }, { prisma }, info) {
-    const commentExists = await prisma.exists.Comment({ id });
+  async updateComment(_parent, { id, data }, { prisma, userId }, info) {
+    if (!userId) throw new Error("Authentication required");
+
+    const commentExists = await prisma.exists.Comment({
+      id,
+      author: { id: userId }
+    });
     if (!commentExists) throw new Error("Comment doesn't exist");
 
     return await prisma.mutation.updateComment({ where: { id }, data }, info);
   },
 
-  async deleteComment(_parent, { id }, { prisma }, info) {
-    const commentExists = await prisma.exists.Comment({ id });
+  async deleteComment(_parent, { id }, { prisma, userId }, info) {
+    if (!userId) throw new Error("Authentication required");
+
+    const commentExists = await prisma.exists.Comment({
+      id,
+      author: { id: userId }
+    });
     if (!commentExists) throw new Error("Comment doesn't exist");
 
     return await prisma.mutation.deleteComment({ where: { id } }, info);
